feat: ini tx query
