+++ apps/app/app/(wallet)/[address]/transactions/(components)/transactions-data-table.tsx
@@ -15,24 +15,15 @@
 
 "use client";
 
-import { getUserOperations, getUserOperationsCount } from "@lightdotso/client";
-import {
-  keepPreviousData,
-  useQuery,
-  useQueryClient,
-} from "@tanstack/react-query";
+import { useQueryClient } from "@tanstack/react-query";
 import { useMemo, type FC } from "react";
 import type { Address } from "viem";
 import { columns } from "@/app/(wallet)/[address]/transactions/(components)/data-table/columns";
 import { DataTable } from "@/app/(wallet)/[address]/transactions/(components)/data-table/data-table";
-import type {
-  UserOperationCountData,
-  UserOperationData,
-  WalletSettingsData,
-} from "@/data";
+import type { WalletSettingsData } from "@/data";
+import { useQueryUserOperations, useQueryUserOperationsCount } from "@/query";
 import { queryKeys } from "@/queryKeys";
 import { usePaginationQueryState } from "@/queryStates";
-import { useAuth } from "@/stores";
 
 // -----------------------------------------------------------------------------
 // Props
@@ -51,12 +42,6 @@ export const TransactionsDataTable: FC<TransactionsDataTableProps> = ({
   address,
   status,
 }) => {
-  // ---------------------------------------------------------------------------
-  // Stores
-  // ---------------------------------------------------------------------------
-
-  const { clientType } = useAuth();
-
   // ---------------------------------------------------------------------------
   // Query State Hooks
   // ---------------------------------------------------------------------------
@@ -80,102 +65,20 @@ export const TransactionsDataTable: FC<TransactionsDataTableProps> = ({
   const walletSettings: WalletSettingsData | undefined =
     queryClient.getQueryData(queryKeys.wallet.settings({ address }).queryKey);
 
-  const currentData: UserOperationData[] | undefined = queryClient.getQueryData(
-    queryKeys.user_operation.list({
-      address,
-      status,
-      order: status === "proposed" ? "asc" : "desc",
-      limit: paginationState.pageSize,
-      offset: offsetCount,
-      is_testnet: walletSettings?.is_enabled_testnet ?? false,
-    }).queryKey,
-  );
-
-  const { data: transactions } = useQuery<UserOperationData[] | null>({
-    placeholderData: keepPreviousData,
-    queryKey: queryKeys.user_operation.list({
-      address,
-      status,
-      order: status === "proposed" ? "asc" : "desc",
-      limit: paginationState.pageSize,
-      offset: offsetCount,
-      is_testnet: walletSettings?.is_enabled_testnet ?? false,
-    }).queryKey,
-    queryFn: async () => {
-      const res = await getUserOperations(
-        {
-          params: {
-            query: {
-              address,
-              status,
-              order: status === "proposed" ? "asc" : "desc",
-              limit: paginationState.pageSize,
-              offset: offsetCount,
-              is_testnet: walletSettings?.is_enabled_testnet ?? false,
-            },
-          },
-        },
-        clientType,
-      );
-
-      // Return if the response is 200
-      return res.match(
-        data => {
-          return data;
-        },
-        _ => {
-          return currentData ?? null;
-        },
-      );
-    },
+  const { userOperations } = useQueryUserOperations({
+    address,
+    status,
+    order: status === "proposed" ? "asc" : "desc",
+    limit: paginationState.pageSize,
+    offset: offsetCount,
+    is_testnet: walletSettings?.is_enabled_testnet ?? false,
   });
 
-  const currentCountData: UserOperationCountData | undefined =
-    queryClient.getQueryData(
-      queryKeys.user_operation.listCount({
-        address: address as Address,
-        status,
-        is_testnet: walletSettings?.is_enabled_testnet ?? false,
-      }).queryKey,
-    );
-
-  const { data: userOperationsCount } = useQuery<UserOperationCountData | null>(
-    {
-      queryKey: queryKeys.user_operation.listCount({
-        address: address as Address,
-        status,
-        is_testnet: walletSettings?.is_enabled_testnet ?? false,
-      }).queryKey,
-      queryFn: async () => {
-        if (!address) {
-          return null;
-        }
-
-        const res = await getUserOperationsCount(
-          {
-            params: {
-              query: {
-                address: address,
-                status: status,
-                is_testnet: walletSettings?.is_enabled_testnet ?? false,
-              },
-            },
-          },
-          clientType,
-        );
-
-        // Return if the response is 200
-        return res.match(
-          data => {
-            return data;
-          },
-          _ => {
-            return currentCountData ?? null;
-          },
-        );
-      },
-    },
-  );
+  const { userOperationsCount } = useQueryUserOperationsCount({
+    address,
+    status,
+    is_testnet: walletSettings?.is_enabled_testnet ?? false,
+  });
 
   // ---------------------------------------------------------------------------
   // Effect Hooks
@@ -194,7 +97,7 @@ export const TransactionsDataTable: FC<TransactionsDataTableProps> = ({
 
   return (
     <DataTable
-      data={transactions ?? []}
+      data={userOperations ?? []}
       address={address}
       columns={columns}
       pageCount={pageCount ?? 0}
