+++ apps/app/src/query/index.ts
@@ -17,9 +17,15 @@ export { useMutationWallet } from "./useMutationWallet";
 export { useMutationWalletSettings } from "./useMutationWalletSettings";
 export { useQueryActivities } from "./useQueryActivities";
 export { useQueryActivitiesCount } from "./useQueryActivitiesCount";
+export { useQueryConfiguration } from "./useQueryConfiguration";
+export { useQueryTokens } from "./useQueryTokens";
+export { useQueryTokensCount } from "./useQueryTokensCount";
 export { useQueryTransactions } from "./useQueryTransactions";
 export { useQueryTransactionsCount } from "./useQueryTransactionsCount";
-export { useQueryConfiguration } from "./useQueryConfiguration";
+export { useQueryUserOperations } from "./useQueryUserOperations";
+export { useQueryUserOperationsCount } from "./useQueryUserOperationsCount";
+export { useQueryWallets } from "./useQueryWallets";
+export { useQueryWalletsCount } from "./useQueryWalletsCount";
 export { useSuspenseQueryAuthSession } from "./useSuspenseQueryAuthSession";
 export { useSuspenseQueryPortfolio } from "./useSuspenseQueryPortfolio";
 export { useSuspenseQueryTokenPrice } from "./useSuspenseQueryTokenPrice";
